Unify Xaml Parameters for Dialogs and Navigation
