+++ Source/Xamarin/Prism.Forms/Properties/AssemblyInfo.cs
@@ -13,3 +13,4 @@
 [assembly: Xamarin.Forms.XmlnsDefinition("http://prismlibrary.com", "Prism.Mvvm")]
 [assembly: Xamarin.Forms.XmlnsDefinition("http://prismlibrary.com", "Prism.Navigation.Xaml")]
 [assembly: Xamarin.Forms.XmlnsDefinition("http://prismlibrary.com", "Prism.Services.Dialogs.Xaml")]
+[assembly: Xamarin.Forms.XmlnsDefinition("http://prismlibrary.com", "Prism.Xaml")]
