+++ Source/Xamarin/Prism.Forms/Services/Dialogs/KnownDialogParameters.cs
@@ -3,5 +3,7 @@
     public static class KnownDialogParameters
     {
         public const string CloseOnBackgroundTapped = "closeOnBackgroundTapped";
+
+        public const string XamlParam = "xamlParam";
     }
 }
