--- Source/Xamarin/Prism.Forms/Services/Dialogs/Xaml/DialogParameter.cs
@@ -1,20 +1,0 @@
-﻿using Xamarin.Forms;
-
-namespace Prism.Services.Dialogs.Xaml
-{
-    public class DialogParameter : BindableObject
-    {
-        public string Key { get; set; }
-
-        public static readonly BindableProperty ValueProperty = 
-            BindableProperty.Create(nameof(Value),
-                typeof(object),
-                typeof(DialogParameter));
-
-        public object Value
-        {
-            get => GetValue(ValueProperty);
-            set => SetValue(ValueProperty, value);
-        }
-    }
-}
